Compile clean on MacOS 26.7

+++ macos/coretext/rasterise_text_impl.cpp
@@ -225,8 +225,8 @@ public:
             CGRectMake(
                 0u,
                 0u,
-                static_cast<size_t>(ceil(bounding_box.size.width)),
-                static_cast<size_t>(ceil(bounding_box.size.height))
+                ceil(bounding_box.size.width),
+                ceil(bounding_box.size.height)
             )
         );
         CGContextSetShouldAntialias(
